Added projoct noof reviews column and subscribe api

+++ src/main/java/com/iii/amirutham/model/INProductAvgRating.java
@@ -1,7 +1,7 @@
 package com.iii.amirutham.model;
 
 public interface INProductAvgRating {
-    Long getTotalReviews();
+    Integer getTotalReviews();
     Long getProductId();
     Long getTotalStars();
     float getRating();
